requirements.txt
discord.py
git+https://github.com/roundar/myjson

